New attribute added, small refactoring
