+++ src/ITCC.HTTP.API.Documentation/ITCC.HTTP.API.Documentation.csproj
@@ -78,6 +78,7 @@
   <ItemGroup>
     <Compile Include="Core\DocGenerator.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="Utils\Wrappers.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="README.md" />
